+++ CHANGES.rst
@@ -8,6 +8,8 @@ Deprecated ``helper.connect``.
 Deprecated ``helper.connect_db`` in favor of functions
 now available in pymongo 3.5.
 
+Added ``helper.get_collection``.
+
 9.1
 ===